backend/requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.42.0
pandas==2.3.3
SQLAlchemy==2.0.41
langchain==1.2.13
langchain-anthropic==1.4.0
langchain-experimental==0.4.1
langchain-community==0.4.1
anthropic==0.86.0
pydantic==2.12.5
openpyxl==3.1.5
python-dotenv==1.2.2
python-multipart==0.0.22
tabulate==0.10.0
psycopg2-binary==2.9.11
python-jose[cryptography]==3.3.0
passlib[pbkdf2]==1.7.4
httpx==0.28.1
